requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
django-filter==21.1
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
